+++ examples/private_service_connect/main.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,37 +20,28 @@
 # If that new version includes provider updates, validation of this
 # example may fail until that is done.
 
-# [START vpc_custom_create]
-module "test-vpc-module" {
+module "private_service_connect" {
+  source                     = "../../modules/private-service-connect"
+  project_id                 = var.project_id
+  network_self_link          = module.simple_vpc.network_self_link
+  private_service_connect_ip = "10.3.0.5"
+  forwarding_rule_target     = "all-apis"
+}
+
+module "simple_vpc" {
   source       = "terraform-google-modules/network/google"
   version      = "~> 4.0.1"
-  project_id   = var.project_id # Replace this with your project ID in quotes
-  network_name = "my-custom-mode-network"
+  project_id   = var.project_id
+  network_name = "my-custom-network"
   mtu          = 1460
 
   subnets = [
     {
-      subnet_name   = "subnet-01"
-      subnet_ip     = "10.10.10.0/24"
-      subnet_region = "us-west1"
-    },
-    {
-      subnet_name           = "subnet-02"
-      subnet_ip             = "10.10.20.0/24"
+      subnet_name           = "my-subnetwork"
+      subnet_ip             = "10.0.0.0/24"
       subnet_region         = "us-west1"
       subnet_private_access = "true"
       subnet_flow_logs      = "true"
-    },
-    {
-      subnet_name               = "subnet-03"
-      subnet_ip                 = "10.10.30.0/24"
-      subnet_region             = "us-west1"
-      subnet_flow_logs          = "true"
-      subnet_flow_logs_interval = "INTERVAL_10_MIN"
-      subnet_flow_logs_sampling = 0.7
-      subnet_flow_logs_metadata = "INCLUDE_ALL_METADATA"
-      subnet_flow_logs_filter   = "false"
     }
   ]
 }
-# [END vpc_custom_create]