+++ examples/private_service_connect/variables.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,7 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">= 0.13.0"
-
-  required_providers {
-    google = {
-      source  = "hashicorp/google"
-      version = "<5.0,>= 2.12"
-    }
-  }
-
-  provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:subnets/v5.2.0"
-  }
+variable "project_id" {
+  description = "Project ID for Private Service Connect."
+  type        = string
 }