+++ modules/subnets/main.tf
@@ -36,11 +36,13 @@ resource "google_compute_subnetwork" "subnetwork" {
       aggregation_interval = lookup(each.value, "subnet_flow_logs_interval", "INTERVAL_5_SEC")
       flow_sampling        = lookup(each.value, "subnet_flow_logs_sampling", "0.5")
       metadata             = lookup(each.value, "subnet_flow_logs_metadata", "INCLUDE_ALL_METADATA")
+      filter_expr          = lookup(each.value, "subnet_flow_logs_filter", "true")
     }] : []
     content {
       aggregation_interval = log_config.value.aggregation_interval
       flow_sampling        = log_config.value.flow_sampling
       metadata             = log_config.value.metadata
+      filter_expr          = log_config.value.filter_expr
     }
   }
   network     = var.network_name
@@ -56,4 +58,7 @@ resource "google_compute_subnetwork" "subnetwork" {
     )) :
     var.secondary_ranges[each.value.subnet_name][i]
   ]
+
+  purpose = lookup(each.value, "purpose", null)
+  role    = lookup(each.value, "role", null)
 }