+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [5.2.0](https://github.com/terraform-google-modules/terraform-google-network/compare/v5.1.0...v5.2.0) (2022-07-28)
+
+
+### Features
+
+* add private service connect module ([#368](https://github.com/terraform-google-modules/terraform-google-network/issues/368)) ([4e90bee](https://github.com/terraform-google-modules/terraform-google-network/commit/4e90bee24bad734ec9a8c99f32902b5904f68796))
+
+## [5.1.0](https://github.com/terraform-google-modules/terraform-google-network/compare/v5.0.0...v5.1.0) (2022-05-16)
+
+
+### Features
+
+* add filter_expr control in subnet log_config ([#360](https://github.com/terraform-google-modules/terraform-google-network/issues/360)) ([5f7e227](https://github.com/terraform-google-modules/terraform-google-network/commit/5f7e22782c8f22e5d4f1a6ed448ae70e91a26b8d))
+
+## [5.0.0](https://github.com/terraform-google-modules/terraform-google-network/compare/v4.1.0...v5.0.0) (2022-02-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* update min TPG versions 3.83 (#349)
+
+### Bug Fixes
+
+* update min TPG versions 3.83 ([#349](https://github.com/terraform-google-modules/terraform-google-network/issues/349)) ([74efa6a](https://github.com/terraform-google-modules/terraform-google-network/commit/74efa6a63ed3f2068c3ecf85fc55a914afcd4be2))
+
+## [4.1.0](https://github.com/terraform-google-modules/terraform-google-network/compare/v4.0.1...v4.1.0) (2022-01-18)
+
+
+### Features
+
+* Promote role and purpose of subnet to GA ([#343](https://github.com/terraform-google-modules/terraform-google-network/issues/343)) ([8b98a36](https://github.com/terraform-google-modules/terraform-google-network/commit/8b98a36eecef8d29c7a2435d1b0809cbd8e1ba6f))
+
 ### [4.0.1](https://www.github.com/terraform-google-modules/terraform-google-network/compare/v4.0.0...v4.0.1) (2021-11-11)